+++ .dotnet/tests/Chat/ChatToolTests.cs
@@ -1,10 +1,8 @@
-﻿using Microsoft.VisualStudio.TestPlatform.CommunicationUtilities;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using OpenAI.Chat;
 using OpenAI.Tests.Utility;
 using System;
 using System.ClientModel;
-using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.Json;
@@ -66,10 +64,10 @@ public partial class ChatToolTests : SyncAsyncTestBase
 
 #pragma warning disable CS0618
     private const string GetFavoriteColorForMonthFunctionName = "get_favorite_color_for_month";
-    private static ChatFunction s_getFavoriteColorForMonthFunction = new ChatFunction(
-        GetFavoriteColorForMonthFunctionName,
-        "gets the caller's favorite color for a given month",
-        BinaryData.FromString("""
+    private static ChatFunction s_getFavoriteColorForMonthFunction = new ChatFunction(GetFavoriteColorForMonthFunctionName)
+    {
+        FunctionDescription = "gets the caller's favorite color for a given month",
+        FunctionParameters = BinaryData.FromString("""
             {
                 "type": "object",
                 "properties": {
@@ -81,7 +79,7 @@ public partial class ChatToolTests : SyncAsyncTestBase
                 "required": [ "month_name" ]
             }
             """)
-    );
+    };
 #pragma warning restore CS0618
 
     private const string GetWeatherForCityToolName = "get_weather_for_city";